refactor(sdk): extract WebSocket support to _client/_websocket.py
First step of splitting cullis_sdk/client.py (3624 lines, ~70 methods,
8 domains) into per-domain mixins under cullis_sdk/_client/.

WebSocket chosen as the pilot because it is the most isolated piece of
the SDK: WebSocketConnection plus the two helper methods (_ws_url and
connect_websocket) have no overlap with auth, sessions, messaging or
the AI gateway.

What moves:

- cullis_sdk/_client/_websocket.py — new module exposing
    * WebSocketConnection (full class)
    * _WebSocketMixin (_ws_url + connect_websocket)
- cullis_sdk/client.py — CullisClient now inherits _WebSocketMixin,
  the two methods and the class definition are removed.
  WebSocketConnection is re-exported for backward compatibility.
- cullis_sdk/__init__.py — WebSocketConnection import path updated to
  the new canonical location.

Public surface is unchanged. All of the following still work:

    from cullis_sdk import CullisClient, WebSocketConnection
    from cullis_sdk.client import WebSocketConnection
    CullisClient(...).connect_websocket()

No logic change — pure code movement.
