+++ cullis_sdk/_client/__init__.py
@@ -1,0 +1,6 @@
+"""Internal split of ``CullisClient`` into per-domain mixins.
+
+The public surface is unchanged — import from :mod:`cullis_sdk` or
+:mod:`cullis_sdk.client`. Modules under this package are implementation
+detail and may change without notice.
+"""
